requirements.txt
amqp==2.2.2
billiard==3.5.0.3
boto==2.48.0
celery==4.1.0
Django==2.0.1
django-celery==3.2.2
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-jinja==2.4.1
django-redis==4.8.0
django-storages==1.6.5
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
memoized-property==1.0.3
psycopg2==2.7.3.2
pytz==2017.3
redis==2.10.6
six==1.11.0
sqlparse==0.2.4
vine==1.1.4
